+++ src/main/java/cart/order/application/OrderCommandService.java
@@ -2,22 +2,28 @@ package cart.order.application;
 
 import cart.cart_item.application.CartItemService;
 import cart.cart_item.application.dto.RemoveCartItemRequest;
+import cart.cart_item.domain.CartItem;
 import cart.coupon.domain.Coupon;
 import cart.member.domain.Member;
 import cart.member_coupon.application.MemberCouponCommandService;
 import cart.order.application.dto.RegisterOrderRequest;
+import cart.order.application.mapper.OrderItemMapper;
 import cart.order.application.mapper.OrderMapper;
 import cart.order.dao.OrderDao;
+import cart.order.dao.OrderItemDao;
 import cart.order.dao.entity.OrderEntity;
+import cart.order.dao.entity.OrderItemEntity;
 import cart.order.domain.Order;
+import cart.order.domain.OrderItem;
 import cart.order.domain.OrderStatus;
+import cart.order.domain.OrderedItems;
 import cart.order.exception.CanNotDeleteNotMyOrderException;
-import cart.order.exception.CanNotDiscountPriceMoreThanTotalPrice;
+import cart.order.exception.CanNotDiscountPriceMoreThanTotalPriceException;
+import cart.order.exception.CanNotOrderNotInCartException;
 import cart.order.exception.NotSameTotalPriceException;
-import cart.order_item.application.OrderItemCommandService;
-import cart.order_item.domain.OrderedItems;
 import cart.value_object.Money;
 import java.util.List;
+import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -26,18 +32,18 @@ import org.springframework.transaction.annotation.Transactional;
 public class OrderCommandService {
 
   private final OrderDao orderDao;
-  private final OrderItemCommandService orderItemCommandService;
+  private final OrderItemDao orderItemDao;
   private final CartItemService cartItemService;
   private final MemberCouponCommandService memberCouponCommandService;
 
   public OrderCommandService(
       final OrderDao orderDao,
-      final OrderItemCommandService orderItemCommandService,
+      final OrderItemDao orderItemDao,
       final CartItemService cartItemService,
       final MemberCouponCommandService memberCouponCommandService
   ) {
     this.orderDao = orderDao;
-    this.orderItemCommandService = orderItemCommandService;
+    this.orderItemDao = orderItemDao;
     this.cartItemService = cartItemService;
     this.memberCouponCommandService = memberCouponCommandService;
   }
@@ -52,11 +58,7 @@ public class OrderCommandService {
 
     final Order order = orderDao.findByOrderId(savedOrderId);
 
-    final OrderedItems orderedItems = orderItemCommandService.registerOrderItem(
-        cartItemIds,
-        order,
-        member
-    );
+    final OrderedItems orderedItems = registerOrderItem(cartItemIds, order, member);
     final Money orderItemsTotalPrice = orderedItems.calculateAllItemPrice();
     validateSameTotalPrice(orderItemsTotalPrice, totalPrice);
 
@@ -72,10 +74,37 @@ public class OrderCommandService {
     return savedOrderId;
   }
 
+  private OrderedItems registerOrderItem(
+      final List<Long> cartItemIds,
+      final Order order,
+      final Member member
+  ) {
+    final List<CartItem> cartItems = cartItemService.findCartItemByCartIds(cartItemIds, member);
+    validateAllCartItemInOrder(cartItemIds, cartItems);
+
+    final List<OrderItemEntity> orderItemEntities = OrderItemMapper.mapToOrderItemEntities(
+        cartItems,
+        order
+    );
+    orderItemDao.save(orderItemEntities);
+
+    final List<OrderItem> orderItems = orderItemDao.findByOrderId(order.getId());
+    return new OrderedItems(orderItems);
+  }
+
+  private void validateAllCartItemInOrder(
+      final List<Long> cartItemIds,
+      final List<CartItem> cartItems
+  ) {
+    if (cartItemIds.size() != cartItems.size()) {
+      throw new CanNotOrderNotInCartException("장바구니에 담지 않은 물품은 주문할 수 없습니다.");
+    }
+  }
+
   private void validateCouponExceedTotalPrice(final OrderedItems orderedItems,
       final Coupon coupon) {
     if (coupon.isExceedDiscountFrom(orderedItems.calculateAllItemPrice())) {
-      throw new CanNotDiscountPriceMoreThanTotalPrice("쿠폰 가격이 전체 가격보다 높으면 사용할 수 있습니다.");
+      throw new CanNotDiscountPriceMoreThanTotalPriceException("쿠폰 가격이 전체 가격보다 높으면 사용할 수 있습니다.");
     }
   }
 
@@ -90,7 +119,12 @@ public class OrderCommandService {
 
     validateOrderOwner(order, member);
 
-    orderItemCommandService.deleteBatch(order);
+    final List<Long> deletedOrderItemIds = orderItemDao.findByOrderId(orderId)
+        .stream()
+        .map(OrderItem::getId)
+        .collect(Collectors.toList());
+
+    orderItemDao.deleteBatch(deletedOrderItemIds);
     orderDao.deleteByOrderId(orderId);
   }
