+++ src/main/java/cart/order/application/dto/OrderItemResponse.java
@@ -1,4 +1,4 @@
-package cart.order_item.application.dto;
+package cart.order.application.dto;
 
 import java.math.BigDecimal;
