+++ src/main/java/cart/order/application/dto/SpecificOrderResponse.java
@@ -1,7 +1,6 @@
 package cart.order.application.dto;
 
 import cart.coupon.application.dto.CouponResponse;
-import cart.order_item.application.dto.OrderItemResponse;
 import java.math.BigDecimal;
 import java.time.ZonedDateTime;
 import java.time.format.DateTimeFormatter;
