+++ src/main/java/cart/order/exception/CanNotDiscountPriceMoreThanTotalPriceException.java
@@ -3,9 +3,9 @@ package cart.order.exception;
 import cart.exception.BusinessException;
 import org.springframework.http.HttpStatus;
 
-public class CanNotDiscountPriceMoreThanTotalPrice extends BusinessException {
+public class CanNotDiscountPriceMoreThanTotalPriceException extends BusinessException {
 
-  public CanNotDiscountPriceMoreThanTotalPrice(final String message) {
+  public CanNotDiscountPriceMoreThanTotalPriceException(final String message) {
     super(message, HttpStatus.BAD_REQUEST);
   }
 }
