+++ src/main/java/cart/order/exception/CanNotOrderNotInCartException.java
@@ -1,0 +1,11 @@
+package cart.order.exception;
+
+import cart.exception.BusinessException;
+import org.springframework.http.HttpStatus;
+
+public class CanNotOrderNotInCartException extends BusinessException {
+
+  public CanNotOrderNotInCartException(final String message) {
+    super(message, HttpStatus.BAD_REQUEST);
+  }
+}
