--- src/main/java/cart/order_item/exception/CanNotOrderNotInCart.java
@@ -1,11 +1,0 @@
-package cart.order_item.exception;
-
-import cart.exception.BusinessException;
-import org.springframework.http.HttpStatus;
-
-public class CanNotOrderNotInCart extends BusinessException {
-
-  public CanNotOrderNotInCart(final String message) {
-    super(message, HttpStatus.BAD_REQUEST);
-  }
-}
